Implementation of undo rm "unrm" for ext4
rm (plus more generally, any unlink-related calls) on an unrm-enabled
system will keep around file metadata (inode, negative dentry) for a set
period of time after the rm is executed, in case we want to be able to
restore those files before the expiration time is up. When a user issues
an unrm ioctl() call to ext4, any recently removed files within the
directory the ioctl() call was issued in will be restored.